+++ test/jest/__mock__/index.js
@@ -2,3 +2,4 @@ import './stripesConfig.mock';
 import './stripes.mock';
 import './intl.mock';
 import './stripesComponents.mock';
+import './stripesCore.mock';