+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change history for ui-checkout
 
+## IN PROGRESS
+
+* Update `@folio/circulation` to `8.0.1`. Refs UICHKOUT-849.
+
 ## [9.0.0](https://github.com/folio-org/ui-checkout/tree/v9.0.0) (2023-02-22)
 [Full Changelog](https://github.com/folio-org/ui-checkout/compare/v8.2.0...v9.0.0)